+++ cmd/neofs-adm/internal/modules/storagecfg/config.go
@@ -38,9 +38,9 @@ control:
 morph:
   dial_timeout: 20s  # timeout for side chain NEO RPC client connection
   cache_ttl: 15s  # use TTL cache for side chain GET operations
-  rpc_endpoint:  # side chain N3 RPC endpoints
+  endpoints:  # side chain N3 RPC endpoints
     {{- range .MorphRPC }}
-    - address: wss://{{.}}/ws{{end}}
+    - wss://{{.}}/ws{{end}}
 {{if not .Relay }}
 storage:
   shard_pool_size: 15  # size of per-shard worker pools used for PUT operations