+++ pkg/innerring/innerring.go
@@ -1019,18 +1019,6 @@ func createListener(ctx context.Context, cli *client.Client, p *chainParams) (ev
 
 func createClient(ctx context.Context, p *chainParams, errChan chan<- error) (*client.Client, error) {
 	endpoints := p.cfg.GetStringSlice(p.name + ".endpoints")
-
-	// deprecated endpoints with priorities
-	section := p.name + ".endpoint.client"
-	for i := 0; ; i++ {
-		addr := p.cfg.GetString(fmt.Sprintf("%s.%d.%s", section, i, "address"))
-		if addr == "" {
-			break
-		}
-
-		endpoints = append(endpoints, addr)
-	}
-
 	if len(endpoints) == 0 {
 		return nil, fmt.Errorf("%s chain client endpoints not provided", p.name)
 	}